Move: SkillKeywordManager to util
